- Fix: Trade guide - Last screen - Not able to click 'I have read' checkbox
