+++ shared/presentation/src/commonMain/kotlin/network/bisq/mobile/presentation/ui/uicases/open_trades/OpenTradeListPresenter.kt
@@ -70,9 +70,7 @@ class OpenTradeListPresenter(
     }
 
     fun onOpenTradeGuide() {
-        disableInteractive()
         navigateTo(Routes.TradeGuideOverview)
-        enableInteractive()
     }
 
     fun onCloseTradeGuideConfirmation() {
